Project: fetch project data on the server Remove `project.fetch` from the Project model. Replace it with @helpers/fetchProjectData, which is imported into `getStaticProps` and only runs on the server. That should remove the API code from the client-side code bundles.
