+++ packages/app-project/src/helpers/fetchProjectData/index.js
@@ -1,0 +1,1 @@
+export { default } from './fetchProjectData'
