+++ packages/app-project/src/helpers/getStaticPageProps/getStaticPageProps.js
@@ -1,10 +1,13 @@
 import notFoundError from '@helpers/notFoundError'
+import fetchProjectData from '@helpers/fetchProjectData'
 import fetchWorkflowsHelper from '@helpers/fetchWorkflowsHelper'
 import initStore from '@stores'
-import { getSnapshot } from 'mobx-state-tree'
+import { applySnapshot, getSnapshot } from 'mobx-state-tree'
 
 export default async function getStaticPageProps({ params, query }) {
   const isServer = true
+  // create a temporary store.
+  // we'll take a snapshot of this later, to pass as a page prop.
   const store = initStore(isServer)
   const { env } = query
 
@@ -12,20 +15,23 @@ export default async function getStaticPageProps({ params, query }) {
     Fetch the project
   */
   if (params.owner && params.project) {
-    const { owner, project } = params
-    const projectSlug = `${owner}/${project}`
-    await store.project.fetch(projectSlug, { env })
-    if (!store.project.id) {
-      return notFoundError(`Project ${owner}/${project} was not found`)
+    const projectSlug = `${params.owner}/${params.project}`
+    const project = await fetchProjectData(projectSlug, { env })
+    if (!project.id) {
+      return notFoundError(`Project ${params.owner}/${params.project} was not found`)
     }
+    applySnapshot(store.project, project)
   }
 
-  /*
-    Validate any workflow URLs
-  */
+  // snapshots don't include computed values, so cache the default workflow ID.
+  const workflowID = store.project.defaultWorkflow
   const { project } = getSnapshot(store)
   const language = project.primary_language
   const { active_workflows, default_workflow } = project.links
+  
+  /*
+    Validate any workflow URLs
+  */
   const workflowExists = active_workflows.includes(params.workflowID)
   if (params.workflowID && !workflowExists) {
     return notFoundError(`Workflow ${params.workflowID} was not found`)
@@ -37,11 +43,9 @@ export default async function getStaticPageProps({ params, query }) {
   const workflows = await fetchWorkflowsHelper(language, active_workflows, default_workflow, env)
   const props = {
     project,
-    isServer,
     workflows
   }
 
-  const workflowID = store.project.defaultWorkflow
   if (workflowID) {
     props.workflowID = workflowID
   }
